% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teom_functions.R
\name{assign_wind_angle}
\alias{assign_wind_angle}
\title{Assign upwind/downwind angles for TEOM pair}
\usage{
assign_wind_angle(df1)
}
\arguments{
\item{df1}{Data frame containing TEOM pairs. Locations or each TEOM in *x* & 
*y* columns, matched pairs assigned group in *dca.group* column, each TEOM  
has unique identifier in *deployment.id* column, "N" or "S" position in 
*position* column.}
}
\value{
Original data frame with *upwind.angle* and *downwind.angle* columns 
addded.
}
\description{
Assign upwind/downwind angles for TEOM pair
}

