% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teom_functions.R
\name{define_event}
\alias{define_event}
\title{Filter hourly data by wind direction.}
\usage{
define_event(df1, locs)
}
\arguments{
\item{df1}{Data frame. Data for teoms under consideration.}

\item{locs}{Data frame. List of teoms with relevant upwind direction.}
}
\value{
Data frame containing only those hours that are in an upwind 
direction for the associated teom.
}
\description{
Filter hourly data by wind direction.
}

