% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{find_missing}
\alias{find_missing}
\title{Identify missing hourly data}
\usage{
find_missing(teom_data)
}
\arguments{
\item{teom_data}{Data frame.}
}
\description{
Identify missing hourly data
}
\examples{
find_missing(teom_data)
}

