% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teom_plot_functions.R
\name{g_legend}
\alias{g_legend}
\title{strip legend from ggplot object}
\usage{
g_legend(a.gplot)
}
\arguments{
\item{a.gplot}{ggplot object.}
}
\value{
A grob of the plot legend
}
\description{
strip legend from ggplot object
}

