% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teom_functions.R
\name{pair_teoms}
\alias{pair_teoms}
\title{Pair up teoms for upwind/downwind analysis}
\usage{
pair_teoms(df1)
}
\arguments{
\item{df1}{Data frame. Teoms under consideration.}
}
\value{
Input data frame with added column listing relevant DCA for paired 
TEOMS.
}
\description{
This function is specific to pairing teoms in TwB2 upwind/downwind analysis.
}

