% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teom_plot_functions.R
\name{plot_rose_image_only}
\alias{plot_rose_image_only}
\title{Create stripped down image of directional rose ggplot object.}
\usage{
plot_rose_image_only(data, value, dir, dirres = 22.5, valuemin = 0,
  valueseq, palette = "YlOrRd", plot.label = "")
}
\arguments{
\item{data}{Data frame.}

\item{value}{Text string. Column name of value variable.}

\item{dir}{Text string. Column name of direction variable.}

\item{dirres}{Numeric. Rose paddle angle width.}

\item{valuemin}{Numeric. Lower value limit. All rows with value < valuemin 
are removed from data.}

\item{valueseq}{Numberic. Bin cutoffs for values.}

\item{palette}{Text string. Name of ColorBrewer palette to use for value 
fill.}

\item{plot.label}{Text string. Will be placed at center of rose.}
}
\value{
Returns the plot as a ggplot object. Paddles lengths (total and fill)
are proportional to percentage of total.
}
\description{
Useful for greating image to be overlaid on map as grob.
}
\examples{
plot_rose(datafile, value='sand_flux', dir='resultant_wd_10m',
          valuemin=low.cutoff)
}

