% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{pull_locations}
\alias{pull_locations}
\title{Pull instrument location data.}
\usage{
pull_locations(deploys)
}
\arguments{
\item{deploys}{Numerical. Vector of deployment ids.}
}
\value{
Data frame.
}
\description{
pull instrument UTM coordinates and deployment description from 
instruments.deployments in AirSci PostgreSQL database, base on deployment_id 
number.
}
\examples{
pull_locations(c(1719, 1718, 1849))
}

