% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{pull_mfile}
\alias{pull_mfile}
\title{Pull wind and pm10 data from the GBUAPCD m-files.}
\usage{
pull_mfile(date1, date2)
}
\arguments{
\item{date1, }{date2 Text string. Date range for which to pull data.}
}
\value{
Data frame.
}
\description{
Pull wind speed and direction from teoms.teom_summary_data in AirSci 
PostgreSQL database. This function pulls only the T7 station (required as 
part of the TwB2 paired TEOM analysis.
*Note: The PM10 data in this table is the analog averaged data transmitted 
via LoggerNet. Although the digitial 5 minute data is preferrable for 
reporting, this data is used as the long turn-around for District collected 
PM10 data makes it unavailable for monthly reports.
}
\examples{
pull_mfile("2016-02-01", "2016-03-01")
}

