% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{pull_pm10}
\alias{pull_pm10}
\title{Pull report-quality PM10 data}
\usage{
pull_pm10(date1, date2, deploys)
}
\arguments{
\item{deploys}{Numerical. Vector of deployment ids.}

\item{x}{A number.}
}
\value{
Data frame.
}
\description{
Pull PM10 data for reporting from AirSci PostgreSQL database.
}
\examples{
pull_pm10(c(1719, 1718, 1849))
}

