% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teom_plot_functions.R
\name{teom_pair_plots}
\alias{teom_pair_plots}
\title{Plot paired TEOMS with wind roses.}
\usage{
teom_pair_plots(teom_locs, df1, dcas)
}
\arguments{
\item{teom_locs}{Data frame. TEOMS with pairs grouping in *dca.group* column.}

\item{df1}{Data frame. Hourly PM10 data.}
}
\value{
3 PNG files of plots in **output** subdirectory of working directory.
}
\description{
Plot paired TEOMS with wind roses.
}

