% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teom_plot_functions.R
\name{teom_pair_png_plots}
\alias{teom_pair_png_plots}
\title{Plot paired TEOMS with wind roses.}
\usage{
teom_pair_png_plots(teom_locs, df1)
}
\arguments{
\item{teom_locs}{Data frame. TEOMS with pairs grouping in *dca.group* column.}

\item{df1}{Data frame. Hourly PM10 data.}
}
\value{
3 PNG files of plots in **output** subdirectory of working directory.
}
\description{
Created multiple 3 PNG files (1 for each TwB2 TEOM pair, showing TwB2 areas, 
locations of TEOMS, and wind roses for time period. Function is specific to 
data as generated in teom_pairs_do.R file.
}

